ChaosMod: Don't run ApplyChangeSetEntry hook as late hook
The hook needs to be applied early on, otherwise it's ineffective
